backend/requirements.txt
Flask
requests
python-dotenv
pdf2image